Creating a sidebar
This project consists of creating a sidebar.
Although not widely used in the most current sites, the sidebar is a technique that is very important for the user's interactivity with the site, because in this sidebar, you can set content that can be of interest to the user. In the example created, the sidebar was used for the user to access the social networks of the owner of the site.